Rename the Performance namespace into Resource
