+++ tests/phpunit/Process/Builder/SubscriberBuilderTest.php
@@ -41,10 +41,10 @@ use Infection\Differ\DiffColorizer;
 use Infection\Event\EventDispatcher;
 use Infection\Logger\LoggerFactory;
 use Infection\Mutant\MetricsCalculator;
-use Infection\Performance\Memory\MemoryFormatter;
-use Infection\Performance\Time\Stopwatch;
-use Infection\Performance\Time\TimeFormatter;
 use Infection\Process\Builder\SubscriberBuilder;
+use Infection\Resource\Memory\MemoryFormatter;
+use Infection\Resource\Time\Stopwatch;
+use Infection\Resource\Time\TimeFormatter;
 use Infection\TestFramework\AbstractTestFrameworkAdapter;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
